backend/requirements.txt
sqlalchemy==2.0.23
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
PyMuPDF==1.23.7
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
numpy==1.26.2